+++ test/test_hpack_integration.py
@@ -5,8 +5,10 @@ long time to run, so they're outside the main test suite, but they need to be
 run before every change to HPACK.
 """
 from hpack.hpack import Decoder, Encoder
+from hpack.struct import HeaderTuple
 from binascii import unhexlify
 from pytest import skip
+
 
 class TestHPACKDecoderIntegration(object):
     def test_can_decode_a_story(self, story):
@@ -24,9 +26,16 @@ class TestHPACKDecoderIntegration(object):
             decoded_headers = d.decode(unhexlify(case['wire']))
 
             # The correct headers are a list of dicts, which is annoying.
-            correct_headers = [(item[0], item[1]) for header in case['headers'] for item in header.items()]
+            correct_headers = [
+                (item[0], item[1])
+                for header in case['headers']
+                for item in header.items()
+            ]
             correct_headers = correct_headers
             assert correct_headers == decoded_headers
+            assert all(
+                isinstance(header, HeaderTuple) for header in decoded_headers
+            )
 
     def test_can_encode_a_story_no_huffman(self, raw_story):
         d = Decoder()
@@ -34,12 +43,19 @@ class TestHPACKDecoderIntegration(object):
 
         for case in raw_story['cases']:
             # The input headers are a list of dicts, which is annoying.
-            input_headers = [(item[0], item[1]) for header in case['headers'] for item in header.items()]
+            input_headers = [
+                (item[0], item[1])
+                for header in case['headers']
+                for item in header.items()
+            ]
 
             encoded = e.encode(input_headers, huffman=False)
             decoded_headers = d.decode(encoded)
 
             assert input_headers == decoded_headers
+            assert all(
+                isinstance(header, HeaderTuple) for header in decoded_headers
+            )
 
     def test_can_encode_a_story_with_huffman(self, raw_story):
         d = Decoder()
@@ -47,7 +63,11 @@ class TestHPACKDecoderIntegration(object):
 
         for case in raw_story['cases']:
             # The input headers are a list of dicts, which is annoying.
-            input_headers = [(item[0], item[1]) for header in case['headers'] for item in header.items()]
+            input_headers = [
+                (item[0], item[1])
+                for header in case['headers']
+                for item in header.items()
+            ]
 
             encoded = e.encode(input_headers, huffman=True)
             decoded_headers = d.decode(encoded)